code-quality:move getUrlListFromArgs
