+++ packages/cli/src/index.ts
@@ -18,7 +18,7 @@
  */
 import { Command } from 'commander';
 import events from 'events';
-import { ensureFile, readFile, writeFile } from 'fs-extra';
+import { ensureFile, writeFile } from 'fs-extra';
 // @ts-ignore Package does not support typescript.
 import Spinnies from 'spinnies';
 import { exec } from 'child_process';
@@ -30,16 +30,19 @@ import {
   generateSummaryDataCSV,
   generateTechnologyCSV,
 } from '@ps-analysis-tool/common';
-import { parseStringPromise } from 'xml2js';
 
 /**
  * Internal dependencies.
  */
 import Utility from './utils/utility';
-import { fetchDictionary } from './utils/fetchCookieDictionary';
 import { analyzeCookiesUrlsInBatches } from './procedures/analyzeCookieUrlsInBatches';
 import { analyzeTechnologiesUrlsInBatches } from './procedures/analyzeTechnologiesUrlsInBatches';
-import { delay, validateArgs } from './utils';
+import {
+  fetchDictionary,
+  delay,
+  getUrlListFromArgs,
+  validateArgs,
+} from './utils';
 import { checkPortInUse } from './utils/checkPortInUse';
 
 events.EventEmitter.defaultMaxListeners = 15;
@@ -85,134 +88,6 @@ const initialize = async () => {
     );
     process.exit(1);
   }
-};
-
-const getUrlListFromArgs = async (
-  url: string,
-  sitemapUrl: string,
-  csvPath: string,
-  sitemapPath: string,
-  // @ts-ignore Package does not support typescript.
-  spinnies
-): Promise<string[]> => {
-  let urls: string[] = [];
-
-  if (url) {
-    urls.push(url);
-  } else if (sitemapUrl) {
-    spinnies.add('sitemap-spinner', {
-      text: 'Parsing Sitemap',
-    });
-
-    try {
-      const _urls = await Utility.getUrlsFromSitemap(sitemapUrl);
-      urls = urls.concat(_urls);
-    } catch (error) {
-      console.log('Error parsing the sitemap file');
-      process.exit(1);
-    }
-
-    spinnies.succeed('sitemap-spinner', {
-      text: 'Done parsing Sitemap',
-    });
-  } else if (csvPath) {
-    spinnies.add('csv-spinner', {
-      text: 'Parsing CSV File',
-    });
-
-    try {
-      const csvString = await readFile(csvPath, 'utf-8');
-
-      const lines = csvString.split('\n');
-
-      const _urls = lines.reduce((acc, line) => {
-        if (line.length === 0) {
-          return acc;
-        } else {
-          return acc.concat(line.split(','));
-        }
-      }, [] as string[]);
-
-      if (_urls.length === 0) {
-        console.log('Provided CSV files has no urls');
-        process.exit(1);
-      }
-      _urls.forEach((_url) => {
-        if (!_url.includes('http')) {
-          console.log(`${_url} is not a valid URL`);
-          process.exit(1);
-        }
-      });
-
-      urls = urls.concat(_urls);
-    } catch (error) {
-      console.log('Error reading the CSV file');
-      process.exit(1);
-    }
-
-    spinnies.succeed('csv-spinner', {
-      text: 'Done parsing CSV file',
-    });
-  } else if (sitemapPath) {
-    spinnies.add('sitemap-spinner', {
-      text: 'Parsing XML File',
-    });
-
-    try {
-      const xmlString = await readFile(sitemapPath, 'utf-8');
-      const data = await parseStringPromise(xmlString as string);
-
-      const isSiteIndex = Boolean(data['siteIndex']);
-      const isSiteMap = Boolean(data['urlset']);
-
-      if (isSiteIndex) {
-        console.log(
-          'Sorry, XML Schema for Sitemap index files is not supported by the tool'
-        );
-        process.exit(1);
-      }
-
-      if (!isSiteMap) {
-        console.log('Provided file is not a valid sitemap');
-        process.exit(1);
-      }
-
-      let _urls: string[] = [];
-
-      try {
-        _urls = data['urlset']['url'].reduce(
-          (acc: string[], { loc }: { loc: string | undefined }) =>
-            loc ? acc.concat(loc) : acc,
-          []
-        );
-      } catch (error) {
-        console.log('Provided XML files has no urls in its urlset');
-        process.exit(1);
-      }
-
-      if (_urls.length === 0) {
-        console.log('Provided XML files has no urls ');
-        process.exit(1);
-      }
-      _urls.forEach((_url) => {
-        if (!_url.includes('http')) {
-          console.log(`${_url} is not a valid URL`);
-          process.exit(1);
-        }
-      });
-
-      urls = urls.concat(_urls);
-    } catch (error) {
-      console.log('Error reading the XML file');
-      process.exit(1);
-    }
-
-    spinnies.succeed('sitemap-spinner', {
-      text: 'Done parsing XML file',
-    });
-  }
-
-  return urls;
 };
 
 const saveResults = async (outDir: string, result: CompleteJson[]) => {
